rename use-current-index to use-pagination
